feat: responsive styles for libraries
